Add files via upload
In this project I have made a dashboard of IPL . It includes match ID , match date ,stadium name and so on . First of all, the dataset is preprocessed and well formed so that it will be easy for the visualization of the data . In addition, the KPI is selected and then measures are created for more accurate information. In conclusion, all the required KPIs are used and well managed dashboard is prepared.